Balance Changes V2
Changes, but smaller changes.
Anything relevant is in the BalanceTweaks Changelog file.

Unlike the previous Balance changes, which were focused on making the weapons feels as destinct as possible.

These changes are instead on finding a sweet spot between their traditional UT99 roles and their use in the doom battlefield.

Hopefully I found a sweet spot. Oh and if any weapon feels 'off' please tell me which one it is so I can tweak it more to your liking.

Have a fine day!